fix a few problems

+++ routes/web.php
@@ -22,45 +22,45 @@ Route::get('/activation/{token}', 'Auth\RegisterController@activation', 'auth.ac
 Route::get('/admin', 'Admin\AdminController@index', 'admin.index');
 
 //admin category routes
-Route::get('/category', 'Admin\CategoryController@index', 'admin.category.index');
-Route::get('/category/archive', 'Admin\CategoryController@archive', 'admin.category.archive');
-Route::get('/category/create', 'Admin\CategoryController@create', 'admin.category.create');
-Route::post('/category/store', 'Admin\CategoryController@store', 'admin.category.store');
-Route::get('/category/edit/{id}', 'Admin\CategoryController@edit', 'admin.category.edit');
-Route::put('/category/update/{id}', 'Admin\CategoryController@update', 'admin.category.update');
-Route::delete('/category/delete/{id}', 'Admin\CategoryController@destroy', 'admin.category.delete');
+Route::get('/admin/category', 'Admin\CategoryController@index', 'admin.category.index');
+Route::get('/admin/category/archive', 'Admin\CategoryController@archive', 'admin.category.archive');
+Route::get('/admin/category/create', 'Admin\CategoryController@create', 'admin.category.create');
+Route::post('/admin/category/store', 'Admin\CategoryController@store', 'admin.category.store');
+Route::get('/admin/category/edit/{id}', 'Admin\CategoryController@edit', 'admin.category.edit');
+Route::put('/admin/category/update/{id}', 'Admin\CategoryController@update', 'admin.category.update');
+Route::delete('/admin/category/delete/{id}', 'Admin\CategoryController@destroy', 'admin.category.delete');
 
 //admin product routes
-Route::get('/product', 'Admin\ProductController@index', 'admin.product.index');
-Route::get('/product/archive', 'Admin\ProductController@archive', 'admin.product.archive');
-Route::get('/product/create', 'Admin\ProductController@create', 'admin.product.create');
-Route::post('/product/store', 'Admin\ProductController@store', 'admin.product.store');
-Route::get('/product/edit/{id}', 'Admin\ProductController@edit', 'admin.product.edit');
-Route::put('/product/update/{id}', 'Admin\ProductController@update', 'admin.product.update');
-Route::delete('/product/delete/{id}', 'Admin\ProductController@destroy', 'admin.product.delete');
+Route::get('/admin/product', 'Admin\ProductController@index', 'admin.product.index');
+Route::get('/admin/product/archive', 'Admin\ProductController@archive', 'admin.product.archive');
+Route::get('/admin/product/create', 'Admin\ProductController@create', 'admin.product.create');
+Route::post('/admin/product/store', 'Admin\ProductController@store', 'admin.product.store');
+Route::get('/admin/product/edit/{id}', 'Admin\ProductController@edit', 'admin.product.edit');
+Route::put('/admin/product/update/{id}', 'Admin\ProductController@update', 'admin.product.update');
+Route::delete('/admin/product/delete/{id}', 'Admin\ProductController@destroy', 'admin.product.delete');
 
 //admin gallery routes
-Route::get('/gallery/archive', 'Admin\GalleryController@archive', 'admin.gallery.archive');
-Route::get('/gallery/create/{id}', 'Admin\GalleryController@create', 'admin.gallery.create');
-Route::post('/gallery/store/{id}', 'Admin\GalleryController@store', 'admin.gallery.store');
-Route::get('/gallery/isFirst/{id}', 'Admin\GalleryController@isFirst', 'admin.gallery.isFirst');
-Route::delete('/gallery/delete/{id}', 'Admin\GalleryController@destroy', 'admin.gallery.delete');
+Route::get('/admin/gallery/archive', 'Admin\GalleryController@archive', 'admin.gallery.archive');
+Route::get('/admin/gallery/create/{id}', 'Admin\GalleryController@create', 'admin.gallery.create');
+Route::post('/admin/gallery/store/{id}', 'Admin\GalleryController@store', 'admin.gallery.store');
+Route::get('/admin/gallery/isFirst/{id}', 'Admin\GalleryController@isFirst', 'admin.gallery.isFirst');
+Route::delete('/admin/gallery/delete/{id}', 'Admin\GalleryController@destroy', 'admin.gallery.delete');
 
 //admin brand routes
-Route::get('/brand', 'Admin\BrandController@index', 'admin.brand.index');
-Route::get('/brand/archive', 'Admin\BrandController@archive', 'admin.brand.archive');
-Route::get('/brand/create', 'Admin\BrandController@create', 'admin.brand.create');
-Route::post('/brand/store', 'Admin\BrandController@store', 'admin.brand.store');
-Route::get('/brand/edit/{id}', 'Admin\BrandController@edit', 'admin.brand.edit');
-Route::put('/brand/update/{id}', 'Admin\BrandController@update', 'admin.brand.update');
-Route::delete('/brand/delete/{id}', 'Admin\BrandController@destroy', 'admin.brand.delete');
+Route::get('/admin/brand', 'Admin\BrandController@index', 'admin.brand.index');
+Route::get('/admin/brand/archive', 'Admin\BrandController@archive', 'admin.brand.archive');
+Route::get('/admin/brand/create', 'Admin\BrandController@create', 'admin.brand.create');
+Route::post('/admin/brand/store', 'Admin\BrandController@store', 'admin.brand.store');
+Route::get('/admin/brand/edit/{id}', 'Admin\BrandController@edit', 'admin.brand.edit');
+Route::put('/admin/brand/update/{id}', 'Admin\BrandController@update', 'admin.brand.update');
+Route::delete('/admin/brand/delete/{id}', 'Admin\BrandController@destroy', 'admin.brand.delete');
 
 //admin discount routes
-Route::get('/discount', 'Admin\DiscountController@index', 'admin.discount.index');
-Route::get('/discount/archive', 'Admin\DiscountController@archive', 'admin.discount.archive');
-Route::get('/discount/create', 'Admin\DiscountController@create', 'admin.discount.create');
-Route::post('/discount/store', 'Admin\DiscountController@store', 'admin.discount.store');
-Route::delete('/discount/delete/{id}', 'Admin\DiscountController@destroy', 'admin.discount.delete');
+Route::get('/admin/discount', 'Admin\DiscountController@index', 'admin.discount.index');
+Route::get('/admin/discount/archive', 'Admin\DiscountController@archive', 'admin.discount.archive');
+Route::get('/admin/discount/create', 'Admin\DiscountController@create', 'admin.discount.create');
+Route::post('/admin/discount/store', 'Admin\DiscountController@store', 'admin.discount.store');
+Route::delete('/admin/discount/delete/{id}', 'Admin\DiscountController@destroy', 'admin.discount.delete');
 
 //admin store routes
 Route::get('/admin/store', 'Admin\StoreController@index', 'admin.store.index');
@@ -70,4 +70,3 @@ Route::post('/admin/store/store', 'Admin\StoreController@store', 'admin.store.st
 Route::get('/admin/store/edit/{id}', 'Admin\StoreController@edit', 'admin.store.edit');
 Route::put('/admin/store/update/{id}', 'Admin\StoreController@update', 'admin.store.update');
 Route::delete('/admin/store/delete/{id}', 'Admin\StoreController@destroy', 'admin.store.delete');
-
